CART-566 ft: hash table to store fault attr
Use a hash table to map fault IDs to fault attributes. Fault IDs don't
need to be sequential. The number of fault IDs is only limited by the
range of uint32_t: UINT32_MAX.
