+++ src/include/gurt/common.h
@@ -78,9 +78,6 @@ extern "C" {
  */
 #define _gurt_gettime(ts) clock_gettime(CLOCK_MONOTONIC, ts)
 
-struct d_fault_attr_t *d_fault_attr_mem;
-int d_fault_id_mem;
-
 /* memory allocating macros */
 
 #define D_CHECK_ALLOC(func, cond, ptr, name, size, count, cname,	\
